response entity en region

+++ src/main/java/com/group3/wineshop/Controllers/RegionController.java
@@ -4,6 +4,7 @@ import com.group3.wineshop.Services.RegionService;
 import com.group3.wineshop.entities.Region;
 import com.group3.wineshop.entities.Winery;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,34 +18,45 @@ public class RegionController {
     private RegionService regionService;
 
     @GetMapping("api/region")
-    public List<Region> mostrarTodos() {
-        return regionService.findAll();
+    public ResponseEntity<List<Region>> mostrarTodos() {
+        return new ResponseEntity(regionService.findAll(), HttpStatus.OK);
     }
 
     @GetMapping("api/region/{id}")
-    public Optional<Region> mostrarRegionId(@PathVariable Integer id){
-        return regionService.findById(id);
+    public ResponseEntity<Optional<Region>> mostrarRegionId(@PathVariable Integer id){
+        if(regionService.findById(id).isPresent())
+            return new ResponseEntity(regionService.findById(id),new HttpHeaders(),HttpStatus.OK);
+        else
+            return new ResponseEntity(regionService.findById(id),new HttpHeaders(),HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("api/region/{id}")
-    public void eliminarporId(@PathVariable Integer id){
-        regionService.deleteById(id);
+    public ResponseEntity eliminarporId(@PathVariable Integer id){
+        if(regionService.findById(id).isPresent()) {
+            regionService.deleteById(id);
+            return new ResponseEntity(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
     }
-
-    /*@PostMapping("api/region")
-    public Region crearRegion(@RequestBody Region region){
-        Region temp = regionService.saveRegion(region);
-        return temp;
-    }*/
 
     @PostMapping("api/region")
     public ResponseEntity<Region> crearRegion(@RequestBody Region region){
-        Region tmp = regionService.saveRegion(region);
-        return new ResponseEntity<Region>(tmp, HttpStatus.CREATED);
+        if(regionService.findById(region.getId()).isEmpty()){
+            regionService.saveRegion(region);
+            return new ResponseEntity(HttpStatus.CREATED);
+        }else{
+            return new ResponseEntity(HttpStatus.ALREADY_REPORTED);
+        }
     }
 
     @PutMapping("/api/region")
-    public void modificarRegion(@RequestBody Region r) {
-        regionService.update(r);
+    public ResponseEntity modificarRegion(@RequestBody Region region) {
+        if(regionService.findById(region.getId()).isPresent()) {
+            regionService.update(region);
+            return new ResponseEntity(HttpStatus.OK);
+        } else {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
     }
 }
